Added preview and cancel button, going to fix a lot fo other issues later

+++ image_properties.py
@@ -16,8 +16,6 @@ class ImageProperties:
     rotation: int=0
     brightness: int=0
     contrast: int=1
-    saturation: int=0
-
+    saturation: float=0
     blur_size: int=0
-
     hue: float=0
